+++ packages/jp.co.hololab.dnn.classification/Scripts/ClassificationModel.cs
@@ -4,6 +4,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using Unity.Sentis;
+using Unity.Collections;
 using HoloLab.DNN.Base;
 
 namespace HoloLab.DNN.Classification
@@ -14,16 +15,13 @@ namespace HoloLab.DNN.Classification
     /// </summary>
     public class ClassificationModel : BaseModel, IDisposable
     {
-        private bool apply_softmax = true;
-
         /// <summary>
         /// create classification model from sentis file
         /// </summary>
         /// <param name="file_path">model file path</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ClassificationModel(string file_path, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(file_path, backend_type, apply_quantize)
+        public ClassificationModel(string file_path, BackendType backend_type = BackendType.GPUCompute)
+            : base(file_path, backend_type)
         {
             Initialize();
         }
@@ -33,9 +31,8 @@ namespace HoloLab.DNN.Classification
         /// </summary>
         /// <param name="stream">model stream</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ClassificationModel(System.IO.Stream stream, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(stream, backend_type, apply_quantize)
+        public ClassificationModel(System.IO.Stream stream, BackendType backend_type = BackendType.GPUCompute)
+            : base(stream, backend_type)
         {
             Initialize();
         }
@@ -45,9 +42,8 @@ namespace HoloLab.DNN.Classification
         /// </summary>
         /// <param name="model_asset">model asset</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ClassificationModel(ModelAsset model_asset, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(model_asset, backend_type, apply_quantize)
+        public ClassificationModel(ModelAsset model_asset, BackendType backend_type = BackendType.GPUCompute)
+            : base(model_asset, backend_type)
         {
             Initialize();
         }
@@ -61,12 +57,17 @@ namespace HoloLab.DNN.Classification
         }
 
         /// <summary>
-        /// set apply softmax to confidencies
+        /// apply softmax to confidencies
         /// </summary>
         /// <param name="apply_softmax">apply softmax</param>
-        public void SetApplySoftmax(bool apply_softmax = true)
+        public void ApplySoftmax()
         {
-            this.apply_softmax = apply_softmax;
+            if (is_quantized)
+            {
+                throw new Exception("[error] please run this function before weights quantization.");
+            }
+
+            SetEditedModel(AddPostProcess());
         }
 
         /// <summary>
@@ -78,29 +79,15 @@ namespace HoloLab.DNN.Classification
         {
             var output_tensors = Predict(image);
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
-            var confidences = TensorFloat.AllocNoData(output_tensor.shape);
-            if (apply_softmax)
-            {
-                backend.Softmax(output_tensor, confidences, -1);
-            }
-            else
-            {
-                backend.MemCopy(output_tensor, confidences);
-            }
+            var confidences_tensor = output_tensor.ReadbackAndClone();
+            var confidences = confidences_tensor.AsReadOnlyNativeArray();
 
-            var index = TensorInt.AllocNoData(new TensorShape(1));
-            backend.ArgMax(confidences, index, -1, false);
+            var score = confidences.Max();
+            var class_id = confidences.IndexOf(score);
 
-            confidences = confidences.ReadbackAndClone();
-            index = index.ReadbackAndClone();
-
-            var class_id = index[0];
-            var score = confidences[class_id];
-
-            confidences.Dispose();
-            index.Dispose();
+            confidences_tensor.Dispose();
             output_tensors.AllDispose();
 
             return (class_id, score);
@@ -117,29 +104,15 @@ namespace HoloLab.DNN.Classification
             var output_tensors = new Dictionary<string, Tensor>();
             yield return CoroutineHandler.StartStaticCoroutine(Predict(image, (outputs) => output_tensors = outputs));
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
-            var confidences = TensorFloat.AllocNoData(output_tensor.shape);
-            if (apply_softmax)
-            {
-                backend.Softmax(output_tensor, confidences, -1);
-            }
-            else
-            {
-                backend.MemCopy(output_tensor, confidences);
-            }
+            var confidences_tensor = output_tensor.ReadbackAndClone();
+            var confidences = confidences_tensor.AsReadOnlyNativeArray();
 
-            var index = TensorInt.AllocNoData(new TensorShape(1));
-            backend.ArgMax(confidences, index, -1, false);
+            var score = confidences.Max();
+            var class_id = confidences.IndexOf(score);
 
-            confidences = confidences.ReadbackAndClone();
-            index = index.ReadbackAndClone();
-
-            var class_id = index[0];
-            var score = confidences[class_id];
-
-            confidences.Dispose();
-            index.Dispose();
+            confidences_tensor.Dispose();
             output_tensors.AllDispose();
 
             return_callback((class_id, score));
@@ -155,35 +128,24 @@ namespace HoloLab.DNN.Classification
         {
             var output_tensors = Predict(image);
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
-            var confidences = TensorFloat.AllocNoData(output_tensor.shape);
-            if (apply_softmax)
-            {
-                backend.Softmax(output_tensor, confidences, -1);
-            }
-            else
-            {
-                backend.MemCopy(output_tensor, confidences);
-            }
+            var confidences_tensor = output_tensor.ReadbackAndClone();
+            var confidences = confidences_tensor.AsReadOnlyNativeArray();
 
-            var topk_values = TensorFloat.AllocNoData(output_tensor.shape);
-            var topk_indices = TensorInt.AllocNoData(output_tensor.shape);
-            backend.TopK(confidences, topk_values, topk_indices, topk, -1, true);
-
-            topk_values = topk_values.ReadbackAndClone();
-            topk_indices = topk_indices.ReadbackAndClone();
+            var sorted_confidences = confidences.Select((confidence, index) => (confidence, index))
+                                                .OrderBy(tuple => -tuple.confidence)
+                                                .ToArray();
 
             var classes = new List<(int, float)>(topk);
             foreach (int i in Enumerable.Range(0, topk))
             {
-                var class_id = topk_indices[i];
-                var score = topk_values[i];
+                var score = sorted_confidences[i].confidence;
+                var class_id = sorted_confidences[i].index;
                 classes.Add((class_id, score));
             }
 
-            topk_values.Dispose();
-            topk_indices.Dispose();
+            confidences_tensor.Dispose();
             output_tensors.AllDispose();
 
             return classes;
@@ -201,35 +163,24 @@ namespace HoloLab.DNN.Classification
             var output_tensors = new Dictionary<string, Tensor>();
             yield return CoroutineHandler.StartStaticCoroutine(Predict(image, (outputs) => output_tensors = outputs));
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
-            var confidences = TensorFloat.AllocNoData(output_tensor.shape);
-            if (apply_softmax)
-            {
-                backend.Softmax(output_tensor, confidences, -1);
-            }
-            else
-            {
-                backend.MemCopy(output_tensor, confidences);
-            }
+            var confidences_tensor = output_tensor.ReadbackAndClone();
+            var confidences = confidences_tensor.AsReadOnlyNativeArray();
 
-            var topk_values = TensorFloat.AllocNoData(output_tensor.shape);
-            var topk_indices = TensorInt.AllocNoData(output_tensor.shape);
-            backend.TopK(confidences, topk_values, topk_indices, topk, -1, true);
-
-            topk_values = topk_values.ReadbackAndClone();
-            topk_indices = topk_indices.ReadbackAndClone();
+            var sorted_confidences = confidences.Select((confidence, index) => (confidence, index))
+                                                .OrderBy(tuple => -tuple.confidence)
+                                                .ToArray();
 
             var classes = new List<(int, float)>(topk);
             foreach (int i in Enumerable.Range(0, topk))
             {
-                var class_id = topk_indices[i];
-                var score = topk_values[i];
+                var score = sorted_confidences[i].confidence;
+                var class_id = sorted_confidences[i].index;
                 classes.Add((class_id, score));
             }
 
-            topk_values.Dispose();
-            topk_indices.Dispose();
+            confidences_tensor.Dispose();
             output_tensors.AllDispose();
 
             return_callback(classes);
@@ -239,7 +190,25 @@ namespace HoloLab.DNN.Classification
         {
             SetInputMean(new Vector3(0.485f, 0.456f, 0.406f));
             SetInputStd(new Vector3(0.229f, 0.224f, 0.225f));
-            SetLayersPerFrame(runtime_model.layers.Count / 5); // TODO : automatic adjust number of layers per frame
+            SetSliceFrames(5); // TODO : automatic adjust number of layers per frame
+        }
+
+        private Model AddPostProcess()
+        {
+            try
+            {
+                var functional_graph = new FunctionalGraph();
+                var inputs = functional_graph.AddInputs(base.runtime_model);
+                var outputs = Functional.Forward(base.runtime_model, inputs);
+                var confidences = Functional.Softmax(outputs[0]);
+                var edited_model = functional_graph.Compile(confidences);
+
+                return edited_model;
+            }
+            catch (Exception e)
+            {
+                throw new Exception($"[error] can not add post process to model for some reason. ({e.Message})");
+            }
         }
     }
 }