+++ packages/jp.co.hololab.dnn.objectdetection/Scripts/ObjectDetectionModel_YOLOX.cs
@@ -24,9 +24,8 @@ namespace HoloLab.DNN.ObjectDetection
         /// </summary>
         /// <param name="file_path">model file path</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ObjectDetectionModel_YOLOX(string file_path, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(file_path, backend_type, apply_quantize)
+        public ObjectDetectionModel_YOLOX(string file_path, BackendType backend_type = BackendType.GPUCompute)
+            : base(file_path, backend_type)
         {
             Initialize();
         }
@@ -36,9 +35,8 @@ namespace HoloLab.DNN.ObjectDetection
         /// </summary>
         /// <param name="stream">model stream</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ObjectDetectionModel_YOLOX(System.IO.Stream stream, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(stream, backend_type, apply_quantize)
+        public ObjectDetectionModel_YOLOX(System.IO.Stream stream, BackendType backend_type = BackendType.GPUCompute)
+            : base(stream, backend_type)
         {
             Initialize();
         }
@@ -48,9 +46,8 @@ namespace HoloLab.DNN.ObjectDetection
         /// </summary>
         /// <param name="model_asset">model asset</param>
         /// <param name="backend_type">backend type for inference engine</param>
-        /// <param name="apply_quantize">apply float16 quantize</param>
-        public ObjectDetectionModel_YOLOX(ModelAsset model_asset, BackendType backend_type = BackendType.GPUCompute, bool apply_quantize = true)
-            : base(model_asset, backend_type, apply_quantize)
+        public ObjectDetectionModel_YOLOX(ModelAsset model_asset, BackendType backend_type = BackendType.GPUCompute)
+            : base(model_asset, backend_type)
         {
             Initialize();
         }
@@ -77,7 +74,7 @@ namespace HoloLab.DNN.ObjectDetection
 
             var output_tensors = Predict(input_texture);
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
             var objects = PostProcess(output_tensor, resize_ratio, score_threshold);
             objects = NonMaximumSuppression.NMS(objects, score_threshold, iou_threshold);
@@ -104,7 +101,7 @@ namespace HoloLab.DNN.ObjectDetection
             var output_tensors = new Dictionary<string, Tensor>();
             yield return CoroutineHandler.StartStaticCoroutine(Predict(input_texture, (outputs) => output_tensors = outputs));
             var output_name = runtime_model.outputs[0].name;
-            var output_tensor = output_tensors[output_name] as TensorFloat;
+            var output_tensor = output_tensors[output_name] as Tensor<float>;
 
             var objects = PostProcess(output_tensor, resize_ratio, score_threshold);
             objects = NonMaximumSuppression.NMS(objects, score_threshold, iou_threshold);
@@ -117,9 +114,6 @@ namespace HoloLab.DNN.ObjectDetection
 
         private void Initialize()
         {
-            SetInputMax(255.0f);
-            SetLayersPerFrame(runtime_model.layers.Count / 5); // TODO : automatic adjust number of layers per frame
-
             input_shape = GetInputShapes().First().Value;
             var input_width = input_shape[3];
             var input_height = input_shape[2];
@@ -127,6 +121,9 @@ namespace HoloLab.DNN.ObjectDetection
             var wsizes = strides.Select(stride => input_width / stride).ToList();
             var hsizes = strides.Select(stride => input_height / stride).ToList();
             (grids, expanded_strides) = CreateGridsAndExpandedStrides(wsizes, hsizes);
+
+            SetInputMax(255.0f);
+            SetSliceFrames(5); // TODO : automatic adjust number of layers per frame
         }
 
         private (List<Vector2Int>, List<int>) CreateGridsAndExpandedStrides(List<int> wsizes, List<int> hsizes)
@@ -170,10 +167,10 @@ namespace HoloLab.DNN.ObjectDetection
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private List<HoloLab.DNN.ObjectDetection.Object> PostProcess(TensorFloat output_tensor, float resize_ratio, float score_threshold = 0.0f)
+        private List<HoloLab.DNN.ObjectDetection.Object> PostProcess(Tensor<float> output_tensor, float resize_ratio, float score_threshold = 0.0f)
         {
             output_tensor = output_tensor.ReadbackAndClone();
-            var output_span = output_tensor.ToReadOnlySpan();
+            var output_span = output_tensor.AsReadOnlySpan();
 
             var objects = new List<HoloLab.DNN.ObjectDetection.Object>();
             var tensor_width = output_tensor.shape[2];